[Refactor] create triton ops module
