+++ vllm_ascend/ops/triton/sigmoid_gating.py
